wave34c: monster_actions/passives test alignment (P0)
Production fix: _lich_cast `dmg = jnp.sum(...)` upcasts int32 -> int64
under JAX_ENABLE_X64, breaking jax.lax.cond branch dtype-equality and
cascading into every monster_special_action() test (7 TypeError failures
on nymph/leprechaun/succubus/dragons/kraken).  Cast the sum back to
int32 (Nethax/nethax/subsystems/monster_actions.py:517).

Test alignment:
- test_green_mold_acid: ERODE_CORRODE is is_primary=False (vendor
  trap.c:225), so corrosion writes to oeroded2, not oeroded.  Updated
  assertion to check oeroded2.
- test_passive_fires_via_melee_attack: combat.melee_attack does not
  call apply_passive_to_player (combat.py:21 defers passive); the
  caller invokes the passive after melee resolution.  Test now
  exercises the two-step sequence directly.

Verified: 31/31 monster_actions_parity + monster_passives_parity pass.

+++ Nethax/nethax/subsystems/monster_actions.py
@@ -514,7 +514,7 @@ def _lich_cast(state, slot: jnp.ndarray, rng: jax.Array):
         sides = _LICH_SPELL_DMG[spell, 1]
         rolls = jax.random.randint(rng_roll, (6,), 1, 7, dtype=jnp.int32)  # up to 6 dice
         mask  = jnp.arange(6) < n
-        dmg   = jnp.sum(jnp.where(mask, jnp.minimum(rolls, jnp.int32(sides)), jnp.int32(0)))
+        dmg   = jnp.sum(jnp.where(mask, jnp.minimum(rolls, jnp.int32(sides)), jnp.int32(0))).astype(jnp.int32)
 
         new_hp = jnp.maximum(
             jnp.int32(0),
